# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenant/players/{id}": {
      "get": {
        "operationId": "players_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      },
      "patch": {
        "operationId": "players_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      },
      "delete": {
        "operationId": "players_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/tenant/players/{id}/account": {
      "get": {
        "operationId": "players_account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      },
      "post": {
        "operationId": "players_bind",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/tenant/players/{id}/skills": {
      "post": {
        "operationId": "players_addSkill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/tenant/players/{id}/skills/{skillId}": {
      "delete": {
        "operationId": "players_removeSkill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skillId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/tenant/players/{id}/availability": {
      "post": {
        "operationId": "players_addAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
    "/api/v1/tenant/players/{id}/availability/{availabilityId}": {
      "delete": {
        "operationId": "players_removeAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availabilityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Players"
        ]
      }
    },
...
  },
  "info": {
    "title": "陪玩门店多租户 SaaS API",
    "description": "一期 API。金额字段一律为十进制字符串分（MoneyFen），禁止 number/浮点。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "tenant/platform access token"
      }
    },
...
  }
}